Anims + Menus + UI
Added some player anims to the code so it is able to activate, however not perfect yet. Made all menus and added them to build. Everything is good to go, just missing the controls and some story context in the options menu. UI finally fixed, was able to turn the textures into sprites and replaced the raw images with regular images.