Result screen audios and auto return to menu
Imported some sound effects to use for the result screens (gameover or win screen). This audio plays when the screen pops up. Also planning to add a chest sound so the player has some feedback on the pressure plate puzzle. Now when the gem is collected, the player will no longer respond to input and will face towards the camera while clapping.