Extended map, added enemy, added puzzle.
Added a second moving enemy, (follows player when in range). To do this I wanted to make the map bigger, to fill up some more space, make it more interesting and more of a puzzle game, I added a timer puzzle. For this I also used a timer sound to round it up.